--- src/utils/getChainData.ts
@@ -1,6 +1,0 @@
-import { ChainData, ChainsData } from "../types";
-
-export const getChainData = (
-  chains: ChainsData,
-  chainId: number | string
-): ChainData | undefined => chains.find(chain => chain.chainId == chainId);
