+++ src/utils/index.ts
@@ -1,4 +1,4 @@
-import { TokenData } from "../types";
+import { ChainData, TokenData } from "../types";
 
 export const getTokenData = (
   tokens: TokenData[],
@@ -9,3 +9,8 @@ export const getTokenData = (
     e =>
       e.address.toLowerCase() === address?.toLowerCase() && e.chainId == chainId
   );
+
+export const getChainData = (
+  chains: ChainData[],
+  chainId: number | string
+): ChainData | undefined => chains.find(chain => chain.chainId == chainId);
